requirements.txt
industrialnext.alpha_camera
opencv-contrib-python==4.7.0.72+industrialnext
tomli
matplotlib
torch==2.0.1+industrialnext
PyYAML
pandas>=1.1.4
psutil
py-cpuinfo
scipy>=1.4.1
seaborn>=0.11.0
torchvision>=0.9.0
tqdm>=4.64.0
requests>=2.23.0
